move client server connections to separate folder
