+++ aether/ae_actions/get_client_cloud_connection.cpp
@@ -20,9 +20,9 @@
 
 #include "aether/client.h"
 #include "aether/server_list/no_filter_server_list_policy.h"
-#include "aether/client_connections/client_to_server_stream.h"
+#include "aether/server_connections/client_to_server_stream.h"
 #include "aether/client_connections/client_cloud_connection.h"
-#include "aether/client_connections/client_connection_manager.h"
+#include "aether/connection_manager/client_connection_manager.h"
 
 #include "aether/ae_actions/ae_actions_tele.h"
