+++ aether/connection_manager/client_server_connection_pool.h
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-#ifndef AETHER_CLIENT_CONNECTIONS_CLIENT_SERVER_CONNECTION_POOL_H_
-#define AETHER_CLIENT_CONNECTIONS_CLIENT_SERVER_CONNECTION_POOL_H_
+#ifndef AETHER_CONNECTION_MANAGER_CLIENT_SERVER_CONNECTION_POOL_H_
+#define AETHER_CONNECTION_MANAGER_CLIENT_SERVER_CONNECTION_POOL_H_
 
 #include <map>
 
 #include "aether/common.h"
 #include "aether/obj/obj_id.h"
 #include "aether/ptr/rc_ptr.h"
-#include "aether/client_connections/client_server_connection.h"
+#include "aether/server_connections/client_server_connection.h"
 
 namespace ae {
 /**
@@ -54,4 +54,4 @@ class ClientServerConnectionPool {
 };
 }  // namespace ae
 
-#endif  // AETHER_CLIENT_CONNECTIONS_CLIENT_SERVER_CONNECTION_POOL_H_
+#endif  // AETHER_CONNECTION_MANAGER_CLIENT_SERVER_CONNECTION_POOL_H_
