+++ aether/server_connections/client_to_server_stream.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef AETHER_CLIENT_CONNECTIONS_CLIENT_TO_SERVER_STREAM_H_
-#define AETHER_CLIENT_CONNECTIONS_CLIENT_TO_SERVER_STREAM_H_
+#ifndef AETHER_SERVER_CONNECTIONS_CLIENT_TO_SERVER_STREAM_H_
+#define AETHER_SERVER_CONNECTIONS_CLIENT_TO_SERVER_STREAM_H_
 
 #include <optional>
 
@@ -93,4 +93,4 @@ class ClientToServerStream final : public ByteIStream {
 };
 }  // namespace ae
 
-#endif  // AETHER_CLIENT_CONNECTIONS_CLIENT_TO_SERVER_STREAM_H_
+#endif  // AETHER_SERVER_CONNECTIONS_CLIENT_TO_SERVER_STREAM_H_
